+++ node/Dockerfile
@@ -9,7 +9,6 @@ COPY package*.json ./
 
 # Instalar las dependencias del proyecto
 RUN npm install
-RUN npm install mysql2 cors bcrypt express body-parser cookie-parser googleapis
 
 # Copiar el resto de los archivos del proyecto
 COPY . .
@@ -18,4 +17,5 @@ COPY . .
 EXPOSE 3000
 
 # Comando para ejecutar el servidor
-CMD ["node", "server.js"]
+CMD ["node", "app.js"]
+
