+++ node/package.json
@@ -1,15 +1,18 @@
 {
     "name": "honeycumb-back",
-    "version": "1.0.0",
-    "description": "Backend para generar y descargar el volcado de la base de datos.",
-    "main": "server.js",
+    "version": "0.4.2",
+    "main": "app.js",
     "scripts": {
-        "start": "node server.js"
+      "start": "node app.js"
     },
-    "author": "Equipo 7",
-    "license": "ISC",
     "dependencies": {
-        "express": "^4.18.2",
-        "mysql": "^2.18.1"
+      "express": "^4.18.2",
+      "mysql2": "^2.3.3",
+      "cors": "^2.8.5",
+      "bcrypt": "^5.0.1",
+      "body-parser": "^1.19.1",
+      "cookie-parser": "^1.4.6",
+      "googleapis": "^105.0.0"
     }
-}
+  }
+
